Change more input code to use int for axes instead of float
This probably breaks TAStudio or something, on the bright side a lot of the
float equality checks are gone. see e12b5d813

+++ src/BizHawk.Client.Common/AutofireController.cs
@@ -47,7 +47,7 @@ namespace BizHawk.Client.Common
 		}
 
 		/// <exception cref="NotImplementedException">always</exception>
-		public float AxisValue(string name)
+		public int AxisValue(string name)
 		{
 			throw new NotImplementedException();
 		}
